git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@193 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/Toolkit.h
@@ -32,7 +32,7 @@ public:
 	bool GetNode2DInfo(ISceneNode* pNode, Node2DInfo* pNode2DInfo);
 
 private:
-	bool To2DScreamPos(vector3df v, position2d<s32>* p);
+	bool To2DScreamPos(vector3df v, position2d<f32>* p);
 
 private:
 	ICameraSceneNode*	m_pCamara;
